remove redundant MovieDetails class
