+++ letterboxdpy/pages/movie_profile.py
@@ -104,6 +104,10 @@ class MovieProfile:
     def get_details(self) -> list:
         return extract_movie_details(self.dom)
 
+    def get_extended_details(self) -> dict:
+        """Get extended details (country, studio, language) from details page."""
+        return extract_movie_extended_details(self.dom)
+
     def get_genres(self) -> list:
         return extract_movie_genres(self.dom, self.slug)
 
@@ -299,6 +303,25 @@ def extract_movie_details(dom):
         )
 
     return details
+
+
+def extract_movie_extended_details(dom) -> dict:
+    """Extract detailed movie information from details page."""
+    dom_details = dom.find("div", {"id": ["tab-details"]})
+
+    data = {"country": [], "studio": [], "language": []}
+
+    if dom_details:
+        for a in dom_details.find_all("a"):
+            text = a.text.strip()
+            if a["href"][1:7] == "studio":
+                data["studio"].append(text)
+            elif a["href"][7:14] == "country":
+                data["country"].append(text)
+            elif a["href"][7:15] == "language":
+                data["language"].append(text)
+
+    return data
 
 
 def extract_movie_genres(dom, slug):
